# openapi.yaml
openapi: 3.0.3
info:
  title: Kopahi API
  version: 1.0.0
  description: |
    REST API for the Kopahi B2B2C agri-marketplace.
    All authenticated endpoints accept a JWT in `Authorization: Bearer <token>`.
servers:
  - url: http://localhost:5000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      parameters:
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
      responses:
        "200": { description: OK }
    post:
      tags: [Products]
      security: [{ bearerAuth: [] }]
      summary: Create product (vendor or admin)
      responses:
        "201": { description: Created }
  /api/products/{id}:
    get:
      tags: [Products]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    put:
      tags: [Products]
      security: [{ bearerAuth: [] }]
      summary: Update product (admin, or vendor on own)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    delete:
      tags: [Products]
      security: [{ bearerAuth: [] }]
      summary: Delete product (admin, or vendor on own)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...
  /api/products/upload:
    post:
      tags: [Products]
      security: [{ bearerAuth: [] }]
      summary: Upload product image (multipart "image")
      responses:
        "200": { description: OK }
...
  /api/wishlist:
    get:
      tags: [Wishlist]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/wishlist/{productId}:
    post:
      tags: [Wishlist]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    delete:
      tags: [Wishlist]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...
  /api/vendor/products:
    get:
      tags: [Vendor]
      security: [{ bearerAuth: [] }]
      summary: Vendor's own products
      responses:
        "200": { description: OK }
...